src/main/java/com/gogo/admin/repository/UserRepository.java
package com.gogo.admin.repository;

import org.springframework.data.jpa.repository.Modifying;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.CrudRepository;
import org.springframework.data.repository.PagingAndSortingRepository;
import org.springframework.data.repository.query.Param;

import com.gogo.admin.domain.User;



//public interface UserRepository extends  CrudRepository<User, Integer>{
public interface UserRepository extends  PagingAndSortingRepository<User, Integer>{
	
	// Select user by email
	@Query("SELECT u FROM User u WHERE u.email = :email")
	public User getUserByEmail(@Param("email") String email);
	
	public Long countById(Integer id);
	
	// Update enabled status
	@Query("UPDATE User u SET u.enabled = ?2 WHERE u.id = ?1")
	@Modifying
	public void updateEnabledStatus(Integer id, boolean enabled);

}
